[refactor] 예약상태 Enum 파일 별도 생성
