+++ src/main/kotlin/kr/hhplus/be/server/enums/ReservationStatus.kt
@@ -1,0 +1,8 @@
+package kr.hhplus.be.server.enums
+
+enum class ReservationStatus {
+    AVAILABLE,
+    PENDING,
+    EXPIRED,
+    CONFIRMED
+}
