+++ src/test/java/kr/hhplus/be/server/application/ReservationServiceTest.kt
@@ -8,8 +8,8 @@ import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
 import kr.hhplus.be.server.domain.model.Reservation
-import kr.hhplus.be.server.domain.model.ReservationStatus
 import kr.hhplus.be.server.domain.repo.ReservationRepo
+import kr.hhplus.be.server.enums.ReservationStatus
 import java.time.LocalDateTime
 
 class ReservationServiceTest : BehaviorSpec({
